Build Miralis as a library
This commit turns Miralis into a library. At a high level it introduces
a new `main.rs` file which contains the entry point (previously in `arch`).
The main file is only expected to be compiler to RISC-V so it is allowed
to contain inline assembly.
The library is intended to be re-used in multiple contexts, including as
part of unit tests and symbolic execution. For that purpose inline
assembly is restricted to `arch/metal.rs`.

This is only a first step in the librarification of Miralis. While this
commit decouples the entry point from the rest if Miralis, the library
is still taking control over much of the machine and execution.
