+++ src/decoder.rs
@@ -775,7 +775,7 @@ mod tests {
     /// https://luplab.gitlab.io/rvcodecjs/
     #[test]
     fn system_instructions() {
-        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() });
+        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() }, 0x100000, 0x2000);
         // ECALL: Environment call.
         assert_eq!(mctx.decode(0x00000073), Instr::Ecall);
         // EBREAK: Environment break.
@@ -805,7 +805,7 @@ mod tests {
 
     #[test]
     fn csr_instructions() {
-        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() });
+        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() }, 0x100000, 0x2000);
 
         // CSRRW: Atomic Read/Write CSR.
         assert_eq!(
@@ -870,7 +870,7 @@ mod tests {
 
     #[test]
     fn access_instructions() {
-        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() });
+        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() }, 0x10000, 0x2000);
 
         assert_eq!(
             mctx.decode(0xff87b703),
@@ -1057,7 +1057,7 @@ mod tests {
 
     #[test]
     fn decode_rd() {
-        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() });
+        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() }, 0x10000, 0x2000);
 
         let base_instruction: usize = 0x30001073;
 
@@ -1112,7 +1112,7 @@ mod tests {
 
     #[test]
     fn decode_rs1() {
-        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() });
+        let mctx = MiralisContext::new(unsafe { Arch::detect_hardware() }, 0x10000, 0x2000);
 
         let base_instruction: usize = 0x30001073;
