+++ src/platform/miralis.rs
@@ -6,14 +6,13 @@ use log::Level;
 use miralis_abi::{failure, miralis_log_fmt, success};
 use spin::Mutex;
 
-use crate::config::{
-    PLATFORM_NB_HARTS, TARGET_FIRMWARE_ADDRESS, TARGET_PAYLOAD_ADDRESS, TARGET_STACK_SIZE,
-};
+use crate::config::{TARGET_FIRMWARE_ADDRESS, TARGET_PAYLOAD_ADDRESS};
 use crate::device::clint::{VirtClint, CLINT_SIZE};
 use crate::device::tester::{VirtTestDevice, TEST_DEVICE_SIZE};
 use crate::device::{self, VirtDevice};
 use crate::driver::ClintDriver;
-use crate::{Platform, _stack_start, _start_address};
+use crate::Platform;
+
 // —————————————————————————— Platform Parameters ——————————————————————————— //
 
 const MIRALIS_START_ADDR: usize = TARGET_FIRMWARE_ADDRESS;
@@ -65,19 +64,8 @@ impl Platform for MiralisPlatform {
         FIRMWARE_START_ADDR
     }
 
-    fn get_miralis_memory_start_and_size() -> (usize, usize) {
-        let size: usize;
-        // SAFETY: The unsafe block is required to get the address of the stack and start of
-        // Miralis, which are external values defined by the linker.
-        // We also ensure that `size` is non-negative and within reasonable bounds
-        unsafe {
-            size = (_stack_start as usize)
-                .checked_sub(_start_address as usize)
-                .and_then(|diff| diff.checked_add(TARGET_STACK_SIZE * PLATFORM_NB_HARTS))
-                .unwrap();
-        }
-
-        (MIRALIS_START_ADDR, size.next_power_of_two())
+    fn get_miralis_start() -> usize {
+        MIRALIS_START_ADDR
     }
 
     fn get_max_valid_address() -> usize {
