+++ src/platform/visionfive2.rs
@@ -8,14 +8,13 @@ use log::Level;
 use spin::Mutex;
 
 use crate::arch::{Arch, Architecture};
-use crate::config::{
-    PLATFORM_NB_HARTS, TARGET_FIRMWARE_ADDRESS, TARGET_STACK_SIZE, TARGET_START_ADDRESS,
-};
+use crate::config::{TARGET_FIRMWARE_ADDRESS, TARGET_START_ADDRESS};
 use crate::device::clint::{VirtClint, CLINT_SIZE};
 use crate::device::tester::{VirtTestDevice, TEST_DEVICE_SIZE};
 use crate::device::{self, VirtDevice};
 use crate::driver::ClintDriver;
-use crate::{Platform, _stack_start, _start_address};
+use crate::Platform;
+
 // —————————————————————————— Platform Parameters ——————————————————————————— //
 
 const SERIAL_PORT_BASE_ADDRESS: usize = 0x10000000;
@@ -82,19 +81,8 @@ impl Platform for VisionFive2Platform {
         FIRMWARE_START_ADDR
     }
 
-    fn get_miralis_memory_start_and_size() -> (usize, usize) {
-        let size: usize;
-        // SAFETY: The unsafe block is required to get the address of the stack and start of
-        // Miralis, which are external values defined by the linker.
-        // We also ensure that `size` is non-negative and within reasonable bounds
-        unsafe {
-            size = (_stack_start as usize)
-                .checked_sub(_start_address as usize)
-                .and_then(|diff| diff.checked_add(TARGET_STACK_SIZE * PLATFORM_NB_HARTS))
-                .unwrap();
-        }
-
-        (MIRALIS_START_ADDR, size.next_power_of_two())
+    fn get_miralis_start() -> usize {
+        MIRALIS_START_ADDR
     }
 
     fn get_max_valid_address() -> usize {
